Move learning activity icon helper
to common utils file. It has nothing to do with Vuetify
icons anymore.
